Store software rx timestamps on received packets
Summary:
In this commit, we're storing the rx timestamp into the `ReceivedUdpPacket.maybeSoftwareTs`. Previously, the only case in which we were storing the timestamp into the aforementioned field was when `shouldUseWrapperRecvmmsgForBatchRecv` is true on the client side. On the other codepaths, we weren't storing it.

Note that `packetReceiveTime`, which is currently used in both the client and the server, is set in userspace by calling `Clock::now()`, which gives the steady clock time. It is not the same as the software rx timestamp.

Reviewed By: mjoras

Differential Revision: D61400513

fbshipit-source-id: 019de7d82496ea916ead14e7eac7961dac0a6df7