Add per packet processing in the read path
Summary: The new path will do socket reading and processing per packet as opposed to reading all the packets available and processing at once. There may be benefits in delaying up to max num packets reads if some late packets arrive while we're processing the data, as opposed to batch reading+processing and yielding the loop for the next read.

Reviewed By: mjoras

Differential Revision: D62278116

fbshipit-source-id: 88d0833f924e63a2d402a81a36eabc80ad764365