Use short timer for loop wake up
Summary:
This is another iteration on D61225298. ev_async_send() works but
apparently it is nnot guaranteed to be delivered after the loop yields - in
some tests we have I see it execute the callbacks in the same loop. Which makes
sense I guess given it simply sends a signal to the loop in a thread safe
manner, but libev documentation is confusing again because it specifically says use this to wake up the loop, so one would think the loop would go to sleep  before it's woken up.
With this short timer the loop does yield and is woken up immediately.

Reviewed By: mjoras

Differential Revision: D62400858

fbshipit-source-id: 9fbc2b92598cc6b98ea806a42c08ddb19d5c4809