Add option to adapt reordering threshold based on inflight packets.
Summary:
This is another attempt to have a more generous reordering threshold to avoid spurious loss. ngtcp2 has had this enabled by default for years.

The idea is basically to adapt it based on inflight packets, so allow up to half the inflight bytes of reordering at a time. This might be a better strategy than our existing code which uses spurious loss as lever to move reordering, since that can end up with cases where the threshold is permanently super high even if there's a period of quiescence.

Reviewed By: kvtsoy

Differential Revision: D64052908

fbshipit-source-id: 7cd80886c1de31a1056da827deb79f4d41313df8