Rework the interactions between the per packet callbacks.
Summary:
Our main use case for this right now is on the client side. So don't waste time doing all the callbacks per packet.

Also make it so the read callbacks don't update the loopers and check for closed streams every packet, defer that to the upper layer. In practice this means that when both options are on, these callbacks (besides the read callbacks) / checks / looper updates only happen once per read event.

Reviewed By: kvtsoy

Differential Revision: D65037206

fbshipit-source-id: 71119c9d2d857fcde747ea68676375d44b52c82f