Add option to use TimerFDs instead of libev timeouts.
Summary:
We think there might be some issues with libev's timers (which uses a heap and the poll timeout). To bypass this on Linux we can use timer fds instead.

This diff introduces a timer fd per timeout. Even for many QUIC connections this should be fine in a client context. From libev's perspective the timers will just be like any other fd with read events.

Reviewed By: kvtsoy

Differential Revision: D64703285

fbshipit-source-id: da817455a9f95b40ca5623ecf617c4c61575902a