Add option to coalesce on insertion to read buffer.
Summary:
This diff adds the ability to coalesce adjoining buffers sometimes on insertion into the read buffer, using the tail room. This only happens in the "rightward" case. E.g. if there's an existing buffer at offset [0, 1200) with 5000 bytes of tail room, and we are inserting a 1200 byte buffer, it will copy it to the tail and the buffer will become [0, 2400).

To make this maximally useful we also have to make it so that the allocated packet buffers are over-allocated instead of packet-sized.

Reviewed By: afrind, kvtsoy

Differential Revision: D65394114

fbshipit-source-id: b92d466baa1d253f0c5854d679e97f14f523c7eb