fix wrong lambda call bug
Summary:
When using msvc and compiling quic with C++20 there's a weird issue that the Looper is initialized with the wrong callbacks.

I suspect that this is because the constructor was creating the lambda as part of the header file.
Moving the constructor code to the cpp file fixed the issue

Reviewed By: hanidamlaj, kvtsoy

Differential Revision: D65994029

fbshipit-source-id: 9c5f47b62d0e0c74b5bce05f926e437e6aa2a3ab