Change semantics of flowControlWindowFrequency to actually be useful
Summary:
I suspect that this was the intention from the start, which is that updates are triggered after windowSize / flowControlWindowFrequency is read.

For the current math this is ONLY true if the factor is two. Thsi is because it is subtracting curReadOffset from curAdvertisedOffset, which is the amount of window _remaining_.

Change the math so it works more usefully and e.g. frequencies of 4 work as expected. Conveniently, the math still works for factors of 2 which is the only currently usable one without crazy side effects.

Reviewed By: jbeshay, kvtsoy

Differential Revision: D66216119

fbshipit-source-id: 6d19855dea1c26440d7cc6f66a8c2a503cae2caf