Do not reset the last write time if not writing in the experimental pacer
Summary:
As title.

This fixes a situation where the experimental pacer would get stuck not sending any packets if the pacing timer triggered very early once. When that happens, the sendBatch would be 0 but the last write time would be updated.

This ensures that if we decide not to write, the lastWrite time is not updated.

Reviewed By: kvtsoy

Differential Revision: D66143272

fbshipit-source-id: 1d470796a19bbe410ec61fcae31c45016904cfb3