fix hangs in watchman oss ubuntu tests by excluding liblzma-dev
Summary:
X-link: https://github.com/facebookincubator/zstrong/pull/1072

watchman oss tests on ubuntu are [hanging 2 hours on github and then failing](https://github.com/facebook/watchman/actions/runs/11989130985/job/33425158244), and locally show errors.

One clue was that tests worked locally on centos stream 9 and fedora 40, pointing to one or more of the ubuntu system packages being the trigger.

Turns out that having the ubuntu xz aka liblzma-dev system packages linked is triggering the issue on both ubuntu 22.04 and 24.04.  Disabled system packages for xz and for glog and libunwind that also bring in the xz system packages on ubuntu.

Why ubuntu's liblzma triggers this problem is unknown at this time.  This change seems to be an improvement in that we get test results, with two tests showing intermittent failutes on ubuntu-22.04 runs.

Reviewed By: ckwalsh

Differential Revision: D66446570

fbshipit-source-id: 8a631e51012c7b90e867268451fd023304039589