Operations to conduct on Stream State when sending a reliable reset
Summary:
We appropriately adjust the following when we send a reliable reset:
- Write buffers, retransmission buffers, loss buffers, etc
- The `sumCurStreamBufferLen`, which seems to indicate the number of bytes sent to the transport minus the number of bytes written out to the wire.

Additionally, I'm introducing a `reliableSizeToPeer` variable within the stream state that keeps track of the reliable size we're sending to a peer. In future, we'll use this variable to ensure that we never increase the reliable size.

Note that none of the added logic is hooked up, so it's not executable.

Reviewed By: mjoras

Differential Revision: D64907994

fbshipit-source-id: 2a372a095c03757949c8866241ba20cabf711333