Make a more generic cancelByteEventCallbacksForStreamInternal
Summary:
With reliable resets, we'll want to cancel byte events where the offset is beyond what needs to be reliably delivered. Right now, we only have the functionality to cancel all byte events, or only those up to a certain point.

I'm adding a more generic `cancelByteEventCallbacksForStreamInternal` that takes in a function that can be used to determine which byte events need to be removed.

Reviewed By: afrind

Differential Revision: D67353621

fbshipit-source-id: fa7e758ee6cd40d247392108c28118b604ad6dbc