Make an isAllDataReceivedUntil function
Summary:
I'm making an `isAllDataReceivedUntil` function that returns true if we've received all data up to and including the specified offset. Note that it doesn't have to be the case that the application has read all of the data. It could be buffered in the QUIC layer, for instance.

The reason I'm making this function is that with reliable resets, we're going to modify when we transition the `StreamRecvState` from `Open` to `Closed`.

Reviewed By: mjoras

Differential Revision: D67658233

fbshipit-source-id: 7c4cdbc8988977efbbd6af1548ea80e6d589661b