RoundRobin
Summary:
This class tracks a group of identifiers and round-robins them out via geNext or peekNext/consume.  There are two ways to advance, either by bytes (passed in) or the number of calls to getNext().

It's implemented using a doubly-linked list, with an optional map for fast deletion of large sets.

Reviewed By: mjoras

Differential Revision: D68641758

fbshipit-source-id: ffb17ef6afee336da0117b07ef4119db867a4881