Ensure transportSettingsOverride fn_ takes precedence
Summary: Previously, we had code that changes the transport settings after the override function. This defeats the purpose of the override function where some overrides would not take place. This changes the function to be called last before setting the transport settings on the transport. Also, to avoid an additional copy of the transportSettings, this changes the function to modify its input transportSettings. The class variable transportSettings_ is a const so it is not possible to pass it by mistake to the override function.

Reviewed By: hanidamlaj, kvtsoy

Differential Revision: D69620347

fbshipit-source-id: fa5b7d4f54de50e4b19b75f284ed3d5e5c681f68