Use app limited signal from last acked packet when advancing state machine
Summary:
Previously, bbr2 used its current app limited state to make decisions on advancing the state machine, this can lead to delayed responses to losses in some cases. This diff changes it to use the appLimited state from the last acked packet.

This diff also refactors how the state related to the current ackEvent is accessed in order to simply the code a little bit.

Reviewed By: mjoras

Differential Revision: D69199424

fbshipit-source-id: e66898cd2347e0d997479b5d50b825c242ef34aa