Improve recovery handling
Summary: This change adds a loss recovery approach similar to the one we're using in bbr1. It reduces the congestion window to inflight bytes and reduces it by lost bytes as long as losses continue. This is a configurable option that can be selectively enabled/disabled for both Startup and non-Startup states of bbr2.

Reviewed By: mjoras

Differential Revision: D69557715

fbshipit-source-id: 12ee20106e57af83a2b4e52caf6254204a91da6f