Import expected-lite
Summary:
We currently make extensive use of folly::Expected. However, this has two downsides:

1. Excessive folly dependencies that we don't really need.
2. No ability to compile without exceptions.

std::expected is currently not available, so in the interim we can import that implements the same interface. This also allows us to compile it without the throwing behaviors all the time in our effort to remove exceptions.

Note two changes were made to the imported file:
1. Always turn off exceptions.
2. Add a temporary hasError() function so the refactor isn't quite so massive. This will be cleaned up in the future.

Eventually when std::expected usage becomes universal, the refactor should amount to a namespace change.

Source is: https://github.com/martinmoene/expected-lite

Reviewed By: kvtsoy

Differential Revision: D73404683

fbshipit-source-id: e5a06b66ec6ca7f65aea12c469eb6c2497e9a458