vendor tiny-optional
Summary:
Previously tiny-optional was not vendored. Since we are now directly exporting it in the API via quic::Optional we need to properly vendor it.

The approach I took was to modify the tiny optional headers directly, as this ensures there is no possible ODR or header problems if someone else in the tree imports it. This will make updating from the upstream more annoying but still a straightforward task.

Reviewed By: sharmafb, kvtsoy

Differential Revision: D74201480

fbshipit-source-id: 110234246dda62cea23a87b047a724a0a993b694