Disable exception throwing in quic::Optional
Summary:
Unilaterally disable throwing (from value()) in quic::Optional. This means that any value() access will be a crash instead of a throw. This is the same behavior as std::optional when -fno-exceptions is used, but we are doing it all the time.

This means that all users of optionals within mvfst and from the API have to explicitly check has_value() in order to avoid a crash.

Reviewed By: kvtsoy

Differential Revision: D74407013

fbshipit-source-id: 12be96e14bdf7052a8a39b54d1d472445e107781