Annotate static initializers with CLANG_LAZY_INIT_TEST
Summary:
Context: Eagerly run static initializers are bad for the performance of our apps, especially when this code is loaded early on. We are trying to make them lazy, either by switching simple strings to constexpr/std::string_view or by using a new compiler optimization that transforms the code according to the construct on first use idiom.

This diff: The quic initializers are taking a long time to run when loading libcoldstart, let's delay them until their first usage using the new compiler optimization.

Reviewed By: yozhu

Differential Revision: D75294086

fbshipit-source-id: fe7c4ff2f5f6284bbb3c7353e6ba12251a1680a0