Do not queue streams for flow control update notifications if closed or reset sent
Summary: If a reset has been sent or the stream has been closed but not reaped yet, the flow control updates are not relevant. If these are left intact, then there is a possibility that checkForClosedStream in the API layer will remove a stream while the callbacks are being processed.

Reviewed By: jbeshay

Differential Revision: D75823348

fbshipit-source-id: 912f19a171337a3e5b72f75af0fabba907629eaf