Fix duplicated initial
Summary:
Meta:

The priming data vector vector ownerhip was not moved, that made us save duplicate packets.

For example: the onPrimingDataAvailable callbacks is currently invoked twice
* 1 - after "writing" the initial
* 2 - after "writing" the first flight of 0-RTT packets

If primingData is not moved correctly the callbacks would look like this:
1. onPrimingDataAvailable([initial])
2. onPrimingDataAvailable([initial, 0RTT packets])

so we would end up with a duplicate initial. This is why the tests were originally written with relaxed checks "CHECK_GE(packets.size(), 2)" instead of "CHECK_EQ(packets.size(),2)"

Reviewed By: mjoras

Differential Revision: D78054738

fbshipit-source-id: 5779d6b0513473c9d37416ca5686726144bd3a37