Fix argument comment in quic/server/test/QuicServerTransportTest.cpp +2
Summary:
Argument comments ensure that the correct values are assigned to the correct arguments.

This diff is a manual fix (assisted by the script in D78190897) of one or more argument comments in quic/server/test/QuicServerTransportTest.cpp. Once existing violations are drawn down we will make mismatches between argument comments and argument names an error.

Reviewed By: hanidamlaj

Differential Revision: D78191328

fbshipit-source-id: eae89d622240b516d23edacdf801583eedcc6b1a