Use consistent packet size for priming connections on IPv4 and IPv6
Summary:
Meta:
Fix for D68804551 comment about handling v4/v6 in QUIC 0-RTT priming.

The issue was that priming packets need to be generated ahead of time without knowing which address family (IPv4 or IPv6) they will be used on later.
Without this change we would set different length for ipv4 and ipv6
- IPv4: 1252 bytes (kDefaultV4UDPSendPacketLen)
- IPv6: 1232 bytes (kDefaultV6UDPSendPacketLen)

For priming connections, this change ensures we always use the smaller, conservative packet size (1232 bytes = kDefaultUDPSendPacketLen) that works on both IPv4 and IPv6 networks.

Used CCA mostly for the tests -- removed most of the text since it's way too verbose

Reviewed By: JunqiWang

Differential Revision: D78899669

fbshipit-source-id: 28efbb388c1b251936b019aab8f69a840de70768