Miscellaneous changes to ContiguousReadCursor for compatibility
Summary: We're going to templatize some functions (for example, `parseLongHeaderInvariant` and `quic::ConnectionId::create`) because those functions are also used outside of the QUIC library, and would still rely on folly::io::Cursor. In order to do so, we'll have to make some of the `ContiguousReadCursor` functions on par with functions in `folly::io::Cursor`.

Reviewed By: jbeshay

Differential Revision: D79822700

fbshipit-source-id: 21efc4cc6073e9be64d535d5bfee6d8f0989a135