Extract TransportErrorCode code to string into a separate file
Summary:
Quic uses Fizz dependencies to parse TLS errors into strings.
When using Quic as part of MNS, it makes more sense to use the same functionality we already have in MNS (at least until we move to Fizz-mobile at some point).

This diff creates the MNS/Mbed based version and still keeps a Fizz based version for non mobile builds.
The MNS version was updated to cover all the same alert values that are present in the Fizz version.

Reviewed By: lnicco

Differential Revision: D79828300

fbshipit-source-id: 4495b267100b90af017d52c1fa1b478ad0fb2687