S558411 HUR build
Summary:
See: S558411
There is a crash on HUR (Horizon Unity Runtime) because of a missing dependency, not packaged into a shared library.
This has started with D79828300, which introduced a dependency on MCF, instead of a dependency on Fizz. MCF is already used in most of the relevant builds and this was authored to reduce binary size.

Although we don't know why the MCF dependency is not packaged as a transitive dependency, this diff will exclude the arvr build from the change in dependencies.

Reviewed By: kvtsoy

Differential Revision: D81927333

fbshipit-source-id: b56619f340a8f82b22c066b5191ba9ff4e49a92e