Introduce QuicPathManager
Summary:
This is a new part of the quic connection state that tracks the status of the available paths. Only one path is active, all others are for probing.

This class works for both the client and the server. It currently only has the basic functionality so it can stay as an independent unit without deeper integration into the transport. Subsequent changes will add more functionality and wire it up to the client and server transports.

Reviewed By: kvtsoy

Differential Revision: D80353122

fbshipit-source-id: 084a41eb434529fc041cd20569d7d7a948e80f19