Fix decrementing of inflight bytes in AckHandlers.cpp
Summary: We were in the process of moving the decrementing of inflight bytes out of the congestion controller. However, we noticed that the number of retransmissions went down a lot. Further investigation showed that we were decrementing inflight bytes only for the ack event, but not for the loss event if it was present in `updateCongestionControllerForAck`.

Reviewed By: kvtsoy

Differential Revision: D82255328

fbshipit-source-id: 0080de7b3db0b09ec438732f1f905ef1cbc2e22a