Handle wildcard peer address in the client to keep the connection state consistent
Summary: The wildcard address is not a valid destination address for sending packets. Linux handles this by sending the packets to localhost. This leaves the client connection state with an inconsistent view where it reads packets from an address different from that it has as its peer. E.g. a client with a peer address of 0.0.0.0:443 will actually be exchanging packets with 127.0.0.0.1. This inconsistent state means that the client cannot discard spurious packets from unknown peer addresses.

Reviewed By: sharmafb

Differential Revision: D83985404

fbshipit-source-id: 77b5bbf3819ed2d5b3da7baba07f4b10d7053c29