Add postwrite method to packet processor
Summary: This diff adds a `postwrite` method that is called at the end of each write loop invocation to allow users to customize writeloops with a before/after type of pattern. This is used for the implementation of request-cost tracking at the protocol layer as tracked in T235176826. This was previously discussed with Joseph Beshay on that task.

Reviewed By: jbeshay

Differential Revision: D84533955

fbshipit-source-id: ae1598040494c2cfc02b225972b39e6d2312bd5f