Eagerly remove unused paths for the client and server
Summary:
Before this change, cleaning up old paths was left to the application on the client side, and for lazy reaping on the server side when all the connection ids are used up.

This change eagerly removes old paths. This reduces the number of paths the client/server has to track. For the client, this also reduces the number of sockets it has to read from.

Reviewed By: kvtsoy

Differential Revision: D85290295

fbshipit-source-id: fac52725b6b93a0caac4902ab97e8d334cf45869