Reduce test flake for QuicClientTransportLiteTest (#411)
Summary:
Pull Request resolved: https://github.com/facebook/mvfst/pull/411

In the client's startCryptoHandshake(), we writeSocketData() then notify onTransportReady() asynchronously on the evb.

If the server responds to the handshake packet fast enough, onTransportReady() would be called after the handshake is completed causing ZeroRtt tests to fail because they expect the one rtt handshake to not have completed yet

This change relaxes the tests for ZeroRTT accept/reject tests.

It also moves up all the expectations to be registered before the client is started and explicitly runs the loop just once.

Together these changes should reduce the flakiness of these tests.

Reviewed By: mjoras

Differential Revision: D85782909

fbshipit-source-id: 1babbdc262bc8a263d7e5ff1391efcfebee53532