Leave the rtt state intact if we don't have a better one
Summary:
The RTT values are used by the PTO timer and the path validation timeout. If they are reset to zero, the timers fallback to default values which cannot accommodate connections with high rtt.

This change avoids resetting the rtt data to 0s if we don't have a cached state to restore.

Reviewed By: sharmafb

Differential Revision: D88191599

fbshipit-source-id: 864a593ff412657acdf293509aa6bd34d00dbbb9